[refactor #41] dto 폴더 구조 변경
