+++ src/main/java/com/project/dogwalker/domain/user/walker/elastic/CustomWalkerSearchRepository.java
@@ -1,6 +1,6 @@
 package com.project.dogwalker.domain.user.walker.elastic;
 
-import com.project.dogwalker.walkersearch.dto.WalkerInfoRequest;
+import com.project.dogwalker.walkersearch.dto.request.WalkerInfoRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
