+++ src/main/java/com/project/dogwalker/domain/user/walker/elastic/CustomWalkerSearchRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.project.dogwalker.domain.user.walker.elastic;
 
 
-import com.project.dogwalker.walkersearch.dto.WalkerInfoRequest;
+import com.project.dogwalker.walkersearch.dto.request.WalkerInfoRequest;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
