+++ src/main/java/com/project/dogwalker/reserve/controller/ReserveController.java
@@ -5,11 +5,11 @@ import com.project.dogwalker.common.resolver.auth.AuthMember;
 import com.project.dogwalker.common.resolver.queryString.QueryStringResolver;
 import com.project.dogwalker.member.dto.MemberInfo;
 import com.project.dogwalker.reserve.dto.ReserveCancel;
-import com.project.dogwalker.reserve.dto.ReserveCheckRequest;
-import com.project.dogwalker.reserve.dto.ReserveListResponse;
-import com.project.dogwalker.reserve.dto.ReserveRequest;
-import com.project.dogwalker.reserve.dto.ReserveResponse;
-import com.project.dogwalker.reserve.dto.ReserveStatusRequest;
+import com.project.dogwalker.reserve.dto.request.ReserveCheckRequest;
+import com.project.dogwalker.reserve.dto.response.ReserveListResponse;
+import com.project.dogwalker.reserve.dto.request.ReserveRequest;
+import com.project.dogwalker.reserve.dto.response.ReserveResponse;
+import com.project.dogwalker.reserve.dto.request.ReserveStatusRequest;
 import com.project.dogwalker.reserve.service.ReserveService;
 import jakarta.validation.Valid;
 import java.util.List;
