+++ src/main/java/com/project/dogwalker/reserve/dto/request/ReserveRequest.java
@@ -1,21 +1,31 @@
-package com.project.dogwalker.reserve.dto;
+package com.project.dogwalker.reserve.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import lombok.ToString;
 
 @Getter
 @Builder
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class ReserveCheckRequest {
+@EqualsAndHashCode
+public class ReserveRequest {
   @NotNull
   private Long walkerId;
   @NotNull
-  private LocalDateTime serviceDate;
+  private LocalDateTime serviceDateTime;
+  @NotNull
+  private Integer timeUnit;
+  @NotNull
+  private Integer price;
+  @NotNull
+  private String payMethod;
 }
