+++ src/main/java/com/project/dogwalker/reserve/service/ReserveService.java
@@ -2,11 +2,11 @@ package com.project.dogwalker.reserve.service;
 
 import com.project.dogwalker.member.dto.MemberInfo;
 import com.project.dogwalker.reserve.dto.ReserveCancel;
-import com.project.dogwalker.reserve.dto.ReserveCheckRequest;
-import com.project.dogwalker.reserve.dto.ReserveListResponse;
-import com.project.dogwalker.reserve.dto.ReserveRequest;
-import com.project.dogwalker.reserve.dto.ReserveResponse;
-import com.project.dogwalker.reserve.dto.ReserveStatusRequest;
+import com.project.dogwalker.reserve.dto.request.ReserveCheckRequest;
+import com.project.dogwalker.reserve.dto.response.ReserveListResponse;
+import com.project.dogwalker.reserve.dto.request.ReserveRequest;
+import com.project.dogwalker.reserve.dto.response.ReserveResponse;
+import com.project.dogwalker.reserve.dto.request.ReserveStatusRequest;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
