+++ src/main/java/com/project/dogwalker/reserve/service/ReserveServiceImpl.java
@@ -26,11 +26,11 @@ import com.project.dogwalker.member.dto.MemberInfo;
 import com.project.dogwalker.notice.dto.NoticeRequest;
 import com.project.dogwalker.notice.service.NoticeService;
 import com.project.dogwalker.reserve.dto.ReserveCancel;
-import com.project.dogwalker.reserve.dto.ReserveCheckRequest;
-import com.project.dogwalker.reserve.dto.ReserveListResponse;
-import com.project.dogwalker.reserve.dto.ReserveRequest;
-import com.project.dogwalker.reserve.dto.ReserveResponse;
-import com.project.dogwalker.reserve.dto.ReserveStatusRequest;
+import com.project.dogwalker.reserve.dto.request.ReserveCheckRequest;
+import com.project.dogwalker.reserve.dto.response.ReserveListResponse;
+import com.project.dogwalker.reserve.dto.request.ReserveRequest;
+import com.project.dogwalker.reserve.dto.response.ReserveResponse;
+import com.project.dogwalker.reserve.dto.request.ReserveStatusRequest;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.HashMap;
