+++ src/main/java/com/project/dogwalker/walkersearch/controller/WalkerInfoController.java
@@ -4,10 +4,10 @@ import com.project.dogwalker.common.resolver.queryString.QueryStringResolver;
 import com.project.dogwalker.common.interceptor.Auth;
 import com.project.dogwalker.common.resolver.auth.AuthMember;
 import com.project.dogwalker.member.dto.MemberInfo;
-import com.project.dogwalker.walkersearch.dto.WalkerInfoResponse;
-import com.project.dogwalker.walkersearch.dto.WalkerInfoRequest;
+import com.project.dogwalker.walkersearch.dto.response.WalkerInfoResponse;
+import com.project.dogwalker.walkersearch.dto.request.WalkerInfoRequest;
 import com.project.dogwalker.walkersearch.dto.WalkerReserveInfo;
-import com.project.dogwalker.walkersearch.dto.WalkerUnAvailDetailResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerUnAvailDetailResponse;
 import com.project.dogwalker.walkersearch.service.WalkerInfoService;
 import jakarta.validation.Valid;
 import java.util.List;
