+++ src/main/java/com/project/dogwalker/walkersearch/dto/response/WalkerTimePriceResponse.java
@@ -1,4 +1,4 @@
-package com.project.dogwalker.walkersearch.dto;
+package com.project.dogwalker.walkersearch.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
