+++ src/main/java/com/project/dogwalker/walkersearch/service/WalkerInfoService.java
@@ -1,10 +1,10 @@
 package com.project.dogwalker.walkersearch.service;
 
 import com.project.dogwalker.member.dto.MemberInfo;
-import com.project.dogwalker.walkersearch.dto.WalkerInfoResponse;
-import com.project.dogwalker.walkersearch.dto.WalkerInfoRequest;
+import com.project.dogwalker.walkersearch.dto.response.WalkerInfoResponse;
+import com.project.dogwalker.walkersearch.dto.request.WalkerInfoRequest;
 import com.project.dogwalker.walkersearch.dto.WalkerReserveInfo;
-import com.project.dogwalker.walkersearch.dto.WalkerUnAvailDetailResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerUnAvailDetailResponse;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
