+++ src/test/java/com/project/dogwalker/domain/user/elastic/CustomWalkerSearchRepositoryImplTest.java
@@ -5,7 +5,7 @@ import com.project.dogwalker.domain.user.User;
 import com.project.dogwalker.domain.user.walker.elastic.WalkerDocument;
 import com.project.dogwalker.domain.user.walker.elastic.WalkerSearchRepository;
 import com.project.dogwalker.support.DomainTest;
-import com.project.dogwalker.walkersearch.dto.WalkerInfoRequest;
+import com.project.dogwalker.walkersearch.dto.request.WalkerInfoRequest;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
