+++ src/test/java/com/project/dogwalker/domain/user/walker/WalkerInfoResponseRepositoryImplTest.java
@@ -8,10 +8,10 @@ import com.project.dogwalker.domain.user.Role;
 import com.project.dogwalker.domain.user.User;
 import com.project.dogwalker.domain.user.UserRepository;
 import com.project.dogwalker.support.DomainTest;
-import com.project.dogwalker.walkersearch.dto.WalkerPermUnAvailDateResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerPermUnAvailDateResponse;
 import com.project.dogwalker.walkersearch.dto.WalkerReserveInfo.Response;
-import com.project.dogwalker.walkersearch.dto.WalkerTempUnAvailDateResponse;
-import com.project.dogwalker.walkersearch.dto.WalkerTimePriceResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerTempUnAvailDateResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerTimePriceResponse;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
