+++ src/test/java/com/project/dogwalker/reserve/service/ReserveServiceImplTest.java
@@ -23,8 +23,8 @@ import com.project.dogwalker.exception.reserve.ReserveException;
 import com.project.dogwalker.member.dto.MemberInfo;
 import com.project.dogwalker.notice.service.NoticeServiceImpl;
 import com.project.dogwalker.reserve.dto.ReserveCancel;
-import com.project.dogwalker.reserve.dto.ReserveCheckRequest;
-import com.project.dogwalker.reserve.dto.ReserveStatusRequest;
+import com.project.dogwalker.reserve.dto.request.ReserveCheckRequest;
+import com.project.dogwalker.reserve.dto.request.ReserveStatusRequest;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
