+++ src/test/java/com/project/dogwalker/walkersearch/service/WalkerInfoResponseServiceImplTest.java
@@ -11,14 +11,14 @@ import com.project.dogwalker.domain.user.UserRepository;
 import com.project.dogwalker.domain.user.walker.elastic.WalkerDocument;
 import com.project.dogwalker.domain.user.walker.elastic.WalkerSearchRepository;
 import com.project.dogwalker.member.dto.MemberInfo;
-import com.project.dogwalker.walkersearch.dto.WalkerInfoResponse;
-import com.project.dogwalker.walkersearch.dto.WalkerInfoRequest;
-import com.project.dogwalker.walkersearch.dto.WalkerPermUnAvailDateResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerInfoResponse;
+import com.project.dogwalker.walkersearch.dto.request.WalkerInfoRequest;
+import com.project.dogwalker.walkersearch.dto.response.WalkerPermUnAvailDateResponse;
 import com.project.dogwalker.walkersearch.dto.WalkerReserveInfo;
 import com.project.dogwalker.walkersearch.dto.WalkerReserveInfo.Response;
-import com.project.dogwalker.walkersearch.dto.WalkerTempUnAvailDateResponse;
-import com.project.dogwalker.walkersearch.dto.WalkerTimePriceResponse;
-import com.project.dogwalker.walkersearch.dto.WalkerUnAvailDetailResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerTempUnAvailDateResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerTimePriceResponse;
+import com.project.dogwalker.walkersearch.dto.response.WalkerUnAvailDetailResponse;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Arrays;
